requirements.txt
Flask==3.1.0
Pillow==11.0.0
pytesseract==0.3.13
deep-translator==1.11.4
